alter path_fuzz

+++ scripts/path_fuzz.sh
@@ -9,9 +9,9 @@ fi
 
 base=$(echo $1 | sed 's/http:\/\///g' | sed 's/https:\/\///g' | sed 's/\///g')
 
-ffuf -u $1/FUZZ -fw 0 -fs 0,1 -ac -mc 200 -w $2 -of json -o "paths/$base"
+ffuf -u $1/FUZZ -fw 0 -mc all -fs 0,1 -ac -w $2 -of json -o "paths/$base"
 
-cat "paths/$base" | jq '.results[].url' >> "paths/${base}_found"
+cat "paths/$base" | jq '.results[] | .url + " Status:" + (.status|tostring) + " Length:" + (.length|tostring) + " Words:" + (.words|tostring)' | sed 's/"//g' >> "paths/${base}_found"
 
 rm "paths/$base"
 
@@ -28,6 +28,7 @@ amount_of_checks=$(($length_of_response/4))
 
 #echo $amount_of_checks
 
+final_grep=""
 for i in $( seq 1 $amount_of_checks )
 do
   echo "counter: $i"
@@ -35,7 +36,7 @@ do
   #echo $size
   offset_word=$(echo $codes | awk -v var="$size" '{print $var}')
 
-  if [ $offset_word -ge 500 ]; then
+  if [ $offset_word -ge 30 ]; then
     begin=$(($size+1))
     end=$(($size*4))
     to_ignore=$(echo $codes | cut -d" " -f${begin}-${end})
@@ -46,10 +47,9 @@ do
     fi
   fi
 
-  echo "grep -v ${final_grep}"
-
-  cat "paths/${base}_found" | grep -v "${final_grep}" >> "paths/${base}_triaged"
+  if [[ ! -z $final_grep ]]; then
+    cat "paths/${base}_found" | grep -v "${final_grep}" >> "paths/${base}_triaged"
+    rm "paths/${base}_found"
+  fi
 
 done
-
-rm "paths/${base}_found"
